--- src/app/overview/page.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const OverviewPage = () => {
-  return (
-    <div>Overview</div>
-  )
-}
-
-export default OverviewPage
