+++ src/components/modals/ReportsModal.tsx
@@ -10,7 +10,7 @@ import {
   DialogTrigger,
 } from "@/components/ui/dialog";
 
-import { Label } from "@/components/ui/label";
+
 import { Button } from "@/components/ui/button";
 import {
   Select,
@@ -19,15 +19,59 @@ import {
   SelectTrigger,
   SelectValue,
 } from "@/components/ui/select";
-import { FileText, Download, Calendar, Filter } from "lucide-react";
+import { FileText, Download,Filter } from "lucide-react";
 import { Separator } from "@/components/ui/separator";
-import { CalendarDateRangePicker } from "@/components/date-range-picker";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { useForm } from "react-hook-form";
+import * as z from "zod";
+import {
+  Form,
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/components/ui/form";
+
+const formSchema = z.object({
+  reportType: z.string({
+    required_error: "Please select a report type.",
+  }),
+  dateRange: z.object({
+    from: z.date(),
+    to: z.date(),
+  }),
+  agent: z.string({
+    required_error: "Please select an agent filter.",
+  }),
+  format: z.string({
+    required_error: "Please select an export format.",
+  }),
+});
 
 interface ReportsModalProps {
   trigger: React.ReactNode;
 }
 
 export function ReportsModal({ trigger }: ReportsModalProps) {
+  const form = useForm<z.infer<typeof formSchema>>({
+    resolver: zodResolver(formSchema),
+    defaultValues: {
+      reportType: "",
+      dateRange: {
+        from: new Date(),
+        to: new Date(),
+      },
+      agent: "",
+      format: "",
+    },
+  });
+
+  function onSubmit(values: z.infer<typeof formSchema>) {
+    console.log(values);
+    // Handle report generation logic here
+  }
+
   return (
     <Dialog>
       <DialogTrigger asChild>
@@ -47,111 +91,126 @@ export function ReportsModal({ trigger }: ReportsModalProps) {
             </div>
           </div>
         </DialogHeader>
-        <div className="grid gap-6 py-4">
-          {/* Report Type Section */}
-          <div className="space-y-4">
-            <div className="flex items-center gap-2 text-sm font-medium text-muted-foreground">
-              <FileText className="h-4 w-4" />
-              Report Type
-            </div>
-            <Separator />
-            <div className="grid gap-4">
-              <div className="grid gap-2">
-                <Label htmlFor="report-type" className="text-sm font-medium">
-                  Select Report
-                </Label>
-                <Select>
-                  <SelectTrigger id="report-type">
-                    <SelectValue placeholder="Choose report type" />
-                  </SelectTrigger>
-                  <SelectContent>
-                    <SelectItem value="call-summary">
-                      <div className="flex items-center gap-2">
-                        <div className="h-2 w-2 rounded-full bg-blue-500" />
-                        Call Summary
-                      </div>
-                    </SelectItem>
-                    <SelectItem value="agent-performance">
-                      <div className="flex items-center gap-2">
-                        <div className="h-2 w-2 rounded-full bg-green-500" />
-                        Agent Performance
-                      </div>
-                    </SelectItem>
-                    <SelectItem value="campaign-analytics">
-                      <div className="flex items-center gap-2">
-                        <div className="h-2 w-2 rounded-full bg-orange-500" />
-                        Campaign Analytics
-                      </div>
-                    </SelectItem>
-                  </SelectContent>
-                </Select>
+        <Form {...form}>
+          <form onSubmit={form.handleSubmit(onSubmit)} className="grid gap-6 py-4">
+            {/* Report Type Section */}
+            <div className="space-y-4">
+              <div className="flex items-center gap-2 text-sm font-medium text-muted-foreground">
+                <FileText className="h-4 w-4" />
+                Report Type
+              </div>
+              <Separator />
+              <div className="grid gap-4">
+                <FormField
+                  control={form.control}
+                  name="reportType"
+                  render={({ field }) => (
+                    <FormItem>
+                      <FormLabel>Select Report</FormLabel>
+                      <Select onValueChange={field.onChange} defaultValue={field.value}>
+                        <FormControl>
+                          <SelectTrigger>
+                            <SelectValue placeholder="Choose report type" />
+                          </SelectTrigger>
+                        </FormControl>
+                        <SelectContent>
+                          <SelectItem value="call-summary">
+                            <div className="flex items-center gap-2">
+                              <div className="h-2 w-2 rounded-full bg-blue-500" />
+                              Call Summary
+                            </div>
+                          </SelectItem>
+                          <SelectItem value="agent-performance">
+                            <div className="flex items-center gap-2">
+                              <div className="h-2 w-2 rounded-full bg-green-500" />
+                              Agent Performance
+                            </div>
+                          </SelectItem>
+                          <SelectItem value="campaign-analytics">
+                            <div className="flex items-center gap-2">
+                              <div className="h-2 w-2 rounded-full bg-orange-500" />
+                              Campaign Analytics
+                            </div>
+                          </SelectItem>
+                        </SelectContent>
+                      </Select>
+                      <FormMessage />
+                    </FormItem>
+                  )}
+                />
               </div>
             </div>
-          </div>
 
-          {/* Date Range Section */}
-          <div className="space-y-4">
-            <div className="flex items-center gap-2 text-sm font-medium text-muted-foreground">
-              <Calendar className="h-4 w-4" />
-              Date Range
-            </div>
-            <Separator />
-            <CalendarDateRangePicker />
-          </div>
+        
 
-          {/* Additional Filters */}
-          <div className="space-y-4">
-            <div className="flex items-center gap-2 text-sm font-medium text-muted-foreground">
-              <Filter className="h-4 w-4" />
-              Additional Filters
-            </div>
-            <Separator />
-            <div className="grid gap-4">
-              <div className="grid gap-2">
-                <Label htmlFor="agent" className="text-sm font-medium">
-                  Agent
-                </Label>
-                <Select>
-                  <SelectTrigger id="agent">
-                    <SelectValue placeholder="Select agent" />
-                  </SelectTrigger>
-                  <SelectContent>
-                    <SelectItem value="all">All Agents</SelectItem>
-                    <SelectItem value="active">Active Agents</SelectItem>
-                    <SelectItem value="inactive">Inactive Agents</SelectItem>
-                  </SelectContent>
-                </Select>
+            {/* Additional Filters */}
+            <div className="space-y-4">
+              <div className="flex items-center gap-2 text-sm font-medium text-muted-foreground">
+                <Filter className="h-4 w-4" />
+                Additional Filters
               </div>
-              <div className="grid gap-2">
-                <Label htmlFor="format" className="text-sm font-medium">
-                  Export Format
-                </Label>
-                <Select>
-                  <SelectTrigger id="format">
-                    <SelectValue placeholder="Select format" />
-                  </SelectTrigger>
-                  <SelectContent>
-                    <SelectItem value="pdf">PDF Document</SelectItem>
-                    <SelectItem value="excel">Excel Spreadsheet</SelectItem>
-                    <SelectItem value="csv">CSV File</SelectItem>
-                  </SelectContent>
-                </Select>
+              <Separator />
+              <div className="grid gap-4">
+                <FormField
+                  control={form.control}
+                  name="agent"
+                  render={({ field }) => (
+                    <FormItem>
+                      <FormLabel>Agent</FormLabel>
+                      <Select onValueChange={field.onChange} defaultValue={field.value}>
+                        <FormControl>
+                          <SelectTrigger>
+                            <SelectValue placeholder="Select agent" />
+                          </SelectTrigger>
+                        </FormControl>
+                        <SelectContent>
+                          <SelectItem value="all">All Agents</SelectItem>
+                          <SelectItem value="active">Active Agents</SelectItem>
+                          <SelectItem value="inactive">Inactive Agents</SelectItem>
+                        </SelectContent>
+                      </Select>
+                      <FormMessage />
+                    </FormItem>
+                  )}
+                />
+                <FormField
+                  control={form.control}
+                  name="format"
+                  render={({ field }) => (
+                    <FormItem>
+                      <FormLabel>Export Format</FormLabel>
+                      <Select onValueChange={field.onChange} defaultValue={field.value}>
+                        <FormControl>
+                          <SelectTrigger>
+                            <SelectValue placeholder="Select format" />
+                          </SelectTrigger>
+                        </FormControl>
+                        <SelectContent>
+                          <SelectItem value="pdf">PDF Document</SelectItem>
+                          <SelectItem value="excel">Excel Spreadsheet</SelectItem>
+                          <SelectItem value="csv">CSV File</SelectItem>
+                        </SelectContent>
+                      </Select>
+                      <FormMessage />
+                    </FormItem>
+                  )}
+                />
               </div>
             </div>
-          </div>
-        </div>
-        <DialogFooter className="gap-2 sm:gap-0">
-          <Button variant="outline">
-            Cancel
-          </Button>
-          <Button 
-            type="submit"
-            className="bg-gradient-to-r from-purple-600 to-indigo-600 hover:from-purple-700 hover:to-indigo-700 text-white gap-2"
-          >
-            <Download className="h-4 w-4" />
-            Generate Report
-          </Button>
-        </DialogFooter>
+            <DialogFooter className="gap-2 sm:gap-0">
+              <Button variant="outline" type="button">
+                Cancel
+              </Button>
+              <Button 
+                type="submit"
+                className="bg-gradient-to-r from-purple-600 to-indigo-600 hover:from-purple-700 hover:to-indigo-700 text-white gap-2"
+              >
+                <Download className="h-4 w-4" />
+                Generate Report
+              </Button>
+            </DialogFooter>
+          </form>
+        </Form>
       </DialogContent>
     </Dialog>
   );
