+++ src/components/settings/SystemSettings.tsx
@@ -1,0 +1,67 @@
+import { Card } from "@/components/ui/card";
+import { Label } from "@/components/ui/label";
+import { Switch } from "@/components/ui/switch";
+import { Separator } from "@/components/ui/separator";
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from "@/components/ui/select";
+import { Settings2 } from "lucide-react";
+
+export function SystemSettings() {
+  return (
+    <Card className="border-0">
+      <div className="p-6">
+        <div className="flex items-center gap-2">
+          <Settings2 className="h-5 w-5" />
+          <h3 className="text-lg font-medium">System Preferences</h3>
+        </div>
+        <p className="text-sm text-muted-foreground mt-2">
+          Configure system-wide settings
+        </p>
+
+        <div className="mt-6 space-y-4">
+          <div className="flex flex-col space-y-2">
+            <Label>Time Zone</Label>
+            <Select defaultValue="est">
+              <SelectTrigger>
+                <SelectValue placeholder="Select timezone" />
+              </SelectTrigger>
+              <SelectContent>
+                <SelectItem value="est">Eastern Time</SelectItem>
+                <SelectItem value="cst">Central Time</SelectItem>
+                <SelectItem value="mst">Mountain Time</SelectItem>
+                <SelectItem value="pst">Pacific Time</SelectItem>
+              </SelectContent>
+            </Select>
+          </div>
+
+          <Separator />
+
+          <div className="flex items-center justify-between">
+            <div className="space-y-0.5">
+              <Label>Auto Refresh</Label>
+              <p className="text-sm text-muted-foreground">
+                Automatically refresh dashboard data
+              </p>
+            </div>
+            <Switch defaultChecked />
+          </div>
+
+          <div className="flex items-center justify-between">
+            <div className="space-y-0.5">
+              <Label>Sound Effects</Label>
+              <p className="text-sm text-muted-foreground">
+                Play sound for notifications
+              </p>
+            </div>
+            <Switch defaultChecked />
+          </div>
+        </div>
+      </div>
+    </Card>
+  );
+}
